glibc-rh1505492-types-11.patch

+++ nscd/mem.c
@@ -423,8 +423,8 @@ gc (struct database_dyn *db)
 
       if (__builtin_expect (debug_level >= 3, 0))
 	dbg_log (_("freed %zu bytes in %s cache"),
-		 db->head->first_free
-		 - ((char *) moves->to + moves->size - db->data),
+		 (size_t) (db->head->first_free
+			   - ((char *) moves->to + moves->size - db->data)),
 		 dbnames[db - dbs]);
 
       /* The byte past the end of the last copied block is the next
